refactor(filter): consolidate stripField into core with depth guard
Two parallel stripField implementations landed during wave-1 merges:
- packages/react/src/filter-utils.ts (WS-E+F) — with MAX_FILTER_DEPTH guard
- packages/core/src/filtering.ts (WS-G+J) — without

Keep the core version (already the import target of the property tests),
add the depth guard + MAX_FILTER_DEPTH export to it, remove the react-
side duplicate, update DataGrid.tsx + filter-utils.test.ts imports.
